Safari-609.1.11.1/LayoutTests/http/tests/security/xssAuditor/full-block-object-tag-expected.txt

CONSOLE MESSAGE: line 4: The XSS Auditor blocked access to 'http://127.0.0.1:8000/security/xssAuditor/resources/echo-intertag.pl?test=/security/xssAuditor/full-block-object-tag.html&enable-full-block=1&q=%3Cobject%20name=%27plugin%27%20type=%27application/x-webkit-test-netscape%27%3E%3Cparam%20name=%27movie%27%20value=%27http://localhost:8000/security/xssAuditor/resources/dummy.swf%27%20/%3E%3C/object%3E' because the source code of a script was found within the request. The server sent an 'X-XSS-Protection' header requesting this behavior.
There should be no content in the iframe below:



--------
Frame: 'frame'
--------
